Teach repeat how to handle stdin and stdout.
Repeat captures stdin into memory as the data is passed to the repeated
command, and this captured stdin is then replayed should the command be
repeated. This makes it possible to embed the repeat  tool  into  shell
pipelines.

Repeat  captures  stdout into memory, and if the command was successful
stdout is passed on to stdout as  normal,  while  if  the  command  was
repeated  stdout  is passed to stderr instead. This ensures that output
is passed to stdout once and once only.